+++ CHANGELOG.md
@@ -57,6 +57,8 @@ Version 2.0.0
 - Added option to set address to bind HTTP(S) port to.
 - Added an overview of the installed plugins to the web interface.
 - Added clusterioctl command to list plugins on the master server.
+- Added Player Auth plugin for logging in to the web interface by proving
+  the ability to log in to a server as a given user.
 
 ### Changes